requirements.txt
google-genai 
dotenv